Conflicts:
	com.palantir.typescript/src/com/palantir/typescript/ResourceDeltaVisitor.java
	com.palantir.typescript/src/com/palantir/typescript/TypeScriptBuilder.java
	com.palantir.typescript/src/com/palantir/typescript/services/language/LanguageService.java

+++ com.palantir.typescript/src/com/palantir/typescript/TypeScriptBuilder.java
@@ -29,12 +29,15 @@ import org.eclipse.core.resources.IResourceDelta;
 import org.eclipse.core.resources.IWorkspace;
 import org.eclipse.core.resources.IWorkspaceRunnable;
 import org.eclipse.core.resources.IncrementalProjectBuilder;
+import org.eclipse.core.resources.ProjectScope;
 import org.eclipse.core.resources.ResourcesPlugin;
 import org.eclipse.core.runtime.CoreException;
 import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.core.runtime.IStatus;
 import org.eclipse.core.runtime.Path;
 import org.eclipse.core.runtime.Status;
+import org.eclipse.core.runtime.preferences.IEclipsePreferences;
+import org.eclipse.core.runtime.preferences.IScopeContext;
 import org.eclipse.jface.preference.IPreferenceStore;
 
 import com.google.common.collect.ImmutableList;
@@ -54,6 +57,8 @@ public final class TypeScriptBuilder extends IncrementalProjectBuilder {
     public static final String ID = "com.palantir.typescript.typeScriptBuilder";
 
     private static final String MARKER_TYPE = "com.palantir.typescript.typeScriptProblem";
+
+    private LanguageService cachedLanguageService;
 
     @Override
     protected IProject[] build(int kind, Map args, IProgressMonitor monitor) throws CoreException {
@@ -76,33 +81,38 @@ public final class TypeScriptBuilder extends IncrementalProjectBuilder {
     protected void clean(IProgressMonitor monitor) throws CoreException {
         checkNotNull(monitor);
 
+        // dispose the language service in case it is out-of-sync
+        if (this.cachedLanguageService != null) {
+            this.cachedLanguageService.dispose();
+            this.cachedLanguageService = null;
+        }
+
         this.clean(BuildPathUtils.getAllSourceFilesDeltas(this.getProject()), monitor);
     }
 
     private void build(List<FileDelta> fileDeltas, IProgressMonitor monitor) throws CoreException {
-        // HACKHACK: create a new language service for each build since it seems to have some incorrect caching behavior
-        // fix is: https://typescript.codeplex.com/SourceControl/changeset/8b1915815ce48b5c17772de750a02a38bb309044
-        LanguageService languageService = new LanguageService(this.getProject());
-        try {
-            this.createMarkers(languageService, monitor);
-
-            // compile the source files if compile-on-save is enabled
-            IPreferenceStore preferenceStore = TypeScriptPlugin.getDefault().getPreferenceStore();
-            if (preferenceStore.getBoolean(IPreferenceConstants.COMPILER_COMPILE_ON_SAVE)) {
-                compile(languageService, fileDeltas, monitor);
-            }
-        } finally {
-            languageService.dispose();
+        // compile the source files if compile-on-save is enabled
+        IPreferenceStore preferenceStore = TypeScriptPlugin.getDefault().getPreferenceStore();
+        if (preferenceStore.getBoolean(IPreferenceConstants.COMPILER_COMPILE_ON_SAVE)) {
+            compile(fileDeltas, monitor);
         }
+
+        // create the markers second to ensure the files are compiled as quickly as possible
+        this.createMarkers(monitor);
     }
 
     private void clean(List<FileDelta> fileDeltas, IProgressMonitor monitor) throws CoreException {
-        // clear the problem markers
-        this.getProject().deleteMarkers(MARKER_TYPE, true, IResource.DEPTH_INFINITE);
+        IProject project = this.getProject();
+        IScopeContext projectScope = new ProjectScope(project);
+        IEclipsePreferences projectPreferences = projectScope.getNode(TypeScriptPlugin.ID);
 
-        // remove the built files if compile-on-save is enabled
+        // clear the problem markers
+        project.deleteMarkers(MARKER_TYPE, true, IResource.DEPTH_INFINITE);
+
+        // remove the built files if compile-on-save is enabled and an output directory is not specified
         IPreferenceStore preferenceStore = TypeScriptPlugin.getDefault().getPreferenceStore();
-        if (preferenceStore.getBoolean(IPreferenceConstants.COMPILER_COMPILE_ON_SAVE)) {
+        if (preferenceStore.getBoolean(IPreferenceConstants.COMPILER_COMPILE_ON_SAVE)
+                && projectPreferences.get(IPreferenceConstants.COMPILER_OUTPUT_DIR_OPTION, null) != null) {
             for (FileDelta fileDelta : fileDeltas) {
                 String fileName = fileDelta.getFileName();
                 ImmutableList<String> builtFiles = getBuiltFiles(fileName);
@@ -111,9 +121,13 @@ public final class TypeScriptBuilder extends IncrementalProjectBuilder {
                     Path path = new Path(builtFile);
                     IFile file = ResourcesPlugin.getWorkspace().getRoot().getFileForLocation(path);
 
-                    file.refreshLocal(IResource.DEPTH_ZERO, monitor);
-                    if (file.exists()) {
-                        file.delete(false, monitor);
+                    // refresh the resource for the file if it is within the workspace
+                    if (file != null) {
+                        file.refreshLocal(IResource.DEPTH_ZERO, monitor);
+
+                        if (file.exists()) {
+                            file.delete(false, monitor);
+                        }
                     }
                 }
             }
@@ -130,12 +144,23 @@ public final class TypeScriptBuilder extends IncrementalProjectBuilder {
         ImmutableList<FileDelta> fileDeltas = ResourceDeltaVisitor.getFileDeltas(delta, project);
 
         if (!fileDeltas.isEmpty()) {
+            this.getLanguageService().updateFiles(fileDeltas);
+
             this.clean(fileDeltas, monitor);
             this.build(fileDeltas, monitor);
         }
     }
 
-    private static void compile(LanguageService languageService, List<FileDelta> fileDeltas, IProgressMonitor monitor) throws CoreException {
+    private LanguageService getLanguageService() {
+        if (this.cachedLanguageService == null) {
+            this.cachedLanguageService = new LanguageService(this.getProject());
+        }
+
+        return this.cachedLanguageService;
+    }
+
+
+    private void compile(List<FileDelta> fileDeltas, IProgressMonitor monitor) throws CoreException {
         for (FileDelta fileDelta : fileDeltas) {
             Delta delta = fileDelta.getDelta();
 
@@ -149,7 +174,7 @@ public final class TypeScriptBuilder extends IncrementalProjectBuilder {
 
                 // compile the file
                 try {
-                    compile(fileName, languageService, monitor);
+                    compile(fileName, monitor);
                 } catch (RuntimeException e) {
                     String errorMessage = "Compilation of '" + fileName + "' failed.";
                     Status status = new Status(IStatus.ERROR, TypeScriptPlugin.ID, errorMessage, e);
@@ -160,7 +185,8 @@ public final class TypeScriptBuilder extends IncrementalProjectBuilder {
         }
     }
 
-    private static void compile(String fileName, LanguageService languageService, IProgressMonitor monitor) throws CoreException {
+    private void compile(String fileName, IProgressMonitor monitor) throws CoreException {
+        LanguageService languageService = this.getLanguageService();
         for (String outputFileName : languageService.getEmitOutput(fileName)) {
             Path path = new Path(outputFileName);
             IFile file = ResourcesPlugin.getWorkspace().getRoot().getFileForLocation(path);
@@ -168,22 +194,28 @@ public final class TypeScriptBuilder extends IncrementalProjectBuilder {
             // refresh the resource for the file if it is within the workspace
             if (file != null) {
                 file.refreshLocal(IResource.DEPTH_ZERO, monitor);
-                file.setDerived(true, monitor);
             }
         }
     }
 
-    private void createMarkers(LanguageService languageService, IProgressMonitor monitor) throws CoreException {
-        final Map<String, List<Diagnostic>> diagnostics = languageService.getAllDiagnostics();
+    private void createMarkers(IProgressMonitor monitor) throws CoreException {
+        // HACKHACK: create a new language service for each build since it seems to have some incorrect caching behavior
+        // fix is: https://typescript.codeplex.com/SourceControl/changeset/8b1915815ce48b5c17772de750a02a38bb309044
+        LanguageService languageService = new LanguageService(this.getProject());
+        try {
+            final Map<String, List<Diagnostic>> diagnostics = languageService.getAllDiagnostics();
 
-        // create the markers within a workspace runnable for greater efficiency
-        IWorkspaceRunnable runnable = new IWorkspaceRunnable() {
-            @Override
-            public void run(IProgressMonitor runnableMonitor) throws CoreException {
-                createMarkers(diagnostics);
-            }
-        };
-        ResourcesPlugin.getWorkspace().run(runnable, this.getProject(), IWorkspace.AVOID_UPDATE, monitor);
+            // create the markers within a workspace runnable for greater efficiency
+            IWorkspaceRunnable runnable = new IWorkspaceRunnable() {
+                @Override
+                public void run(IProgressMonitor runnableMonitor) throws CoreException {
+                    createMarkers(diagnostics);
+                }
+            };
+            ResourcesPlugin.getWorkspace().run(runnable, this.getProject(), IWorkspace.AVOID_UPDATE, monitor);
+        } finally {
+            languageService.dispose();
+        }
     }
 
     private static void createMarkers(final Map<String, List<Diagnostic>> diagnostics) throws CoreException {
